rename profile DTO file
